Integrate ssr (#157)
* Enable SSR with migration script

* Add zone.js to polyfills

* Make Map component SSR compatible

* Make "Scroll demo into view" client only

* afterNextRender in layout-toolbar-menu

* afterNextRender in control.component.ts - fixes terrain example

* Lint - Update projects/showcase/src/app/app.config.ts


* use withFetch to fix Error retrieving icon

* add custom loader for svg, fix loading icons on server

* fix loading map for ssr app by removing base-href

* detach portal in browser only

* remove polyfills

* replace window functions

* create popup in browser only with afterNextRender

* cleanup unused lifecycle hooks, remove control only in browser

* prerender with routes file

* Update projects/showcase/src/app/demo/examples/live-update-feature.component.ts

---------

+++ projects/ngx-maplibre-gl/src/lib/control/control.component.ts
@@ -1,14 +1,16 @@
 import {
-  AfterContentInit,
   ChangeDetectionStrategy,
   Component,
   ElementRef,
   Input,
   OnDestroy,
   ViewChild,
+  afterNextRender,
+  inject,
 } from '@angular/core';
 import { ControlPosition, IControl } from 'maplibre-gl';
 import { MapService } from '../map/map.service';
+import { Platform } from '@angular/cdk/platform';
 
 export class CustomControl implements IControl {
   constructor(private container: HTMLElement) {}
@@ -29,9 +31,9 @@ export class CustomControl implements IControl {
 /**
  * `mgl-control` - a custom control component
  * @see [Controls](https://maplibre.org/maplibre-gl-js/docs/API/interfaces/IControl/)
- * 
+ *
  * @category Components
- * 
+ *
  * @example
  * ```html
  * ...
@@ -52,12 +54,12 @@ export class CustomControl implements IControl {
  */
 @Component({
   selector: 'mgl-control',
-  template: '<div class="maplibregl-ctrl" #content><ng-content></ng-content></div>',
+  template:
+    '<div class="maplibregl-ctrl" #content><ng-content></ng-content></div>',
   changeDetection: ChangeDetectionStrategy.OnPush,
   standalone: true,
 })
-export class ControlComponent<T extends IControl>
-  implements OnDestroy, AfterContentInit {
+export class ControlComponent<T extends IControl> implements OnDestroy {
   /** Init input */
   @Input() position?: 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right';
   /** @hidden */
@@ -65,20 +67,24 @@ export class ControlComponent<T extends IControl>
 
   control: T | CustomControl;
 
-  constructor(private mapService: MapService) {}
+  private _isBrowser = inject(Platform).isBrowser;
 
-  ngAfterContentInit() {
-    if (this.content.nativeElement.childNodes.length) {
-      this.control = new CustomControl(this.content.nativeElement);
-      this.mapService.mapCreated$.subscribe(() => {
-        this.mapService.addControl(this.control!, this.position);
-      });
-    }
+  constructor(private mapService: MapService) {
+    afterNextRender(() => {
+      if (this.content.nativeElement.childNodes.length) {
+        this.control = new CustomControl(this.content.nativeElement);
+        this.mapService.mapCreated$.subscribe(() => {
+          this.mapService.addControl(this.control!, this.position);
+        });
+      }
+    });
   }
 
   ngOnDestroy() {
-    if (this.mapService.mapInstance.hasControl(this.control)) {
-      this.mapService.removeControl(this.control);
+    if (this._isBrowser) {
+      if (this.mapService.mapInstance.hasControl(this.control)) {
+        this.mapService.removeControl(this.control);
+      }
     }
   }
 }
